Enable session cache by default
A session cache can reduce the number of Fleet API calls made by
clients, but needed to be explicitly enabled by the user. With this
change, the cache will be enabled by default and saved to
~/.tesla-cache.json. This path can be overridden using the
`-session-cache` command-line flag or by setting TESLA_CACHE_FILE, as
before.

Users can add the `-disable-session-cache` command-line flag to disable
the feature.